FPGA/python_scripts/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
fxpmath==0.4.8
kiwisolver==1.4.4
matplotlib==3.7.1
mpmath==1.3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.1
pick==2.2.0
Pillow==9.5.0
pyadic==0.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sympy==1.12
